requirements.txt
google-adk==1.7.0
yfinance==0.2.65
psutil==7.0.0
litellm==1.74.6
google-generativeai==0.8.5
python-dotenv==1.1.1
protobuf~=5.29.5
pydantic~=2.11.7
uuid==1.30